fix sequencer to use update and dspTime isntead of Time.time. created a CSoundManager and separated some logic

+++ Assets/NoteRecord.cs
@@ -6,12 +6,18 @@ public class NoteRecord
 {
     public float Note { get; set; }
     public float Velocity { get; set; }
-    public float Duration { get; set; }
-    public float StartTime { get; set; }
+    /// <summary>
+    /// In beats, need to be multiplied with BPM to get in seconds
+    /// </summary>
+    public double Duration { get; set; }
+    /// <summary>
+    /// In beats, needs to be multiplied with BPM to get in seconds
+    /// </summary>
+    public double StartTime { get; set; }
 
     public float ReverbAmount { get; set; }
 
-    public NoteRecord(float note, float velocity, float duration = -1f, float startTime = 0f, float reverbAmount = 0f)
+    public NoteRecord(float note, float velocity, double duration = -1f, double startTime = 0f, float reverbAmount = 0f)
     {
         Note = note;
         Velocity = velocity;
